+++ triplea/service/repository/convert_db.py
@@ -1,14 +1,34 @@
 # from triplea.config.settings import SETTINGS,DB_ROOT_PATH
+from typing import Optional
 from triplea.db.mongodb import DB_MongoDB
 from triplea.db.tinydb import DB_TinyDB
 from triplea.schemas.article import Article
 from triplea.service.click_logger import logger
 
 
+def convert_mongodb_to_tinydb(move_oreginal_article: Optional[bool] = False):
+    source_db = DB_MongoDB()
+    article_count = source_db.get_all_article_count()
+    if article_count> 10000:
+        pass
+        # raise NotImplementedError
+
+    la = source_db.get_article_by_state(0)
+    destination_db = DB_TinyDB()
+    n = 0  
+    for a in la:
+        pmid = a['PMID']
+        if destination_db.is_article_exist_by_pmid(pmid):
+            logger.INFO(f'The article {pmid} already exists.', deep = 3 )
+        else:
+            n = n +1
+            if move_oreginal_article == False:
+                a['OreginalArticle'] = {}
+            destination_db.add_new_article(a)
+            logger.DEBUG(f'{n} Copy article {pmid} to destination repository.', deep = 3 )
 
 
-
-def convert():
+def convert_tinydb_to_mongodb():
     source_db = DB_TinyDB()
     article_count = source_db.get_all_article_count()
     if article_count> 10000: