+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 ## v0.0.1 - 2023-02-05
 
 ### Improvements
+- Add Visualization InteractiveGraph
+- Add convert_mongodb_to_tinydb
 - Restructuring project
 - Add Topic Extraction
 - Improve environment variables